refactor: Move internal utils to separate files
