+++ src/state/optimizer-parallel/modes/heuristics.ts
@@ -1,5 +1,5 @@
-import { RUNNING_HEURISTICS } from '../../optimizer/status';
 import type { ExtrasCombinationEntry } from '../../optimizer/types/optimizerSetupTypes';
+import { RUNNING_HEURISTICS } from '../../optimizer/utils/status';
 import { changeStatus } from '../../slices/controlsSlice';
 import type { AppDispatch, RootState } from '../../store';
 import type { WorkerWrapper } from '../calculate';
