+++ src/state/optimizer/utils/combinatorics.ts
