+++ src/state/optimizer/utils/detectFirefox.ts
